refactor out cmd into separate files
